Substantial Refactor
This started as a much smaller effort to fix the authorization check for the `save` hook, since I'm pretty sure that wasn't working. Then I discovered that the `vm` variable name meant at least 3 different things, so I renamed that in a bunch of places. Then I just went for it.

Here are the major changes:
- Fixing the `save` hook so it should work now
- Adding the ability to specify that a type (or all) authorization is mandatory (and causing an error if an authLevel can't be determined).
- Adding a new way to specific/determine the authLevel. This involves using a method on the schema that can take in a specific document and a payload from the query to determine a specific authLevel. The main use case is when the authLevel depends on the document (if you want to check if a specific user owns a specific document, in which case you'd return 'OWNER' as the authLevel, otherwise 'VIEWER'). Since this requires the document, it is done as a post-save hook when doing a find query and it doesn't work for updates/deletes.

I then did a couple passes and simplifying the logic.

Lots of testing will be needed here, but I wanted to share this sooner rather than later to see if this had a shot at being upstreamed and to get some feedback.