Tweaking the asynchronous support
The main change here is to configure the code so it only calls `resolveAuthLevel` once per document. This gets rid of the need for caching.

Also, to handle promises correctly, I bumped the min version of mongoose to 5.x, which makes the code easier to reason about (no callback functions!)